add comms and some folder fixes
