+++ app/src/main/java/yandex/school/project/FinanceApplication.kt
@@ -3,5 +3,9 @@ package yandex.school.project
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
 
+/**
+ * Главный класс приложения, отвечающий за инициализацию Hilt для dependency injection.
+ * Единственная ответственность: настройка и запуск системы dependency injection.
+ */
 @HiltAndroidApp
 class FinanceApplication : Application()
