+++ app/src/main/java/yandex/school/project/MainActivity.kt
@@ -8,6 +8,10 @@ import dagger.hilt.android.AndroidEntryPoint
 import yandex.school.project.presentation.navigation.AppNavigation
 import yandex.school.project.presentation.theme.ProjectTheme
 
+/**
+ * Главная активность приложения, отвечающая за запуск UI и навигации.
+ * Единственная ответственность: инициализация и отображение основного интерфейса приложения.
+ */
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
