+++ app/src/main/java/yandex/school/project/domain/repositories/TransactionRepository.kt
@@ -1,6 +1,7 @@
 package yandex.school.project.domain.repositories
 
 import yandex.school.project.domain.entities.Transaction
+import yandex.school.project.domain.entities.TransactionType
 
 interface TransactionRepository {
     suspend fun createTransaction(
@@ -9,7 +10,7 @@ interface TransactionRepository {
         amount: Double,
         description: String?,
         date: String,
-        type: yandex.school.project.domain.entities.TransactionType
+        type: TransactionType
     ): Transaction
     
     suspend fun getTransactionById(id: Int): Transaction
@@ -20,7 +21,7 @@ interface TransactionRepository {
         amount: Double,
         description: String?,
         date: String,
-        type: yandex.school.project.domain.entities.TransactionType
+        type: TransactionType
     ): Transaction
     
     suspend fun deleteTransaction(id: Int)
