+++ app/src/main/java/yandex/school/project/domain/usecases/account/GetAccountByIdUseCase.kt
@@ -4,6 +4,10 @@ import yandex.school.project.domain.entities.Account
 import yandex.school.project.domain.repositories.AccountRepository
 import javax.inject.Inject
 
+/**
+ * Use case для получения аккаунта по его идентификатору.
+ * Единственная ответственность: получение конкретного аккаунта по ID из репозитория.
+ */
 class GetAccountByIdUseCase @Inject constructor(
     private val accountRepository: AccountRepository
 ) {
