+++ app/src/main/java/yandex/school/project/domain/usecases/account/GetFirstAccountUseCase.kt
@@ -4,6 +4,10 @@ import yandex.school.project.domain.entities.Account
 import yandex.school.project.domain.repositories.AccountRepository
 import javax.inject.Inject
 
+/**
+ * Use case для получения первого доступного аккаунта.
+ * Единственная ответственность: получение первого аккаунта из списка всех аккаунтов.
+ */
 class GetFirstAccountUseCase @Inject constructor(
     private val accountRepository: AccountRepository
 ) {
