+++ app/src/main/java/yandex/school/project/domain/usecases/category/GetCategoriesByTypeUseCase.kt
@@ -4,6 +4,10 @@ import yandex.school.project.domain.entities.Category
 import yandex.school.project.domain.repositories.CategoryRepository
 import javax.inject.Inject
 
+/**
+ * Use case для получения категорий по типу (доходы/расходы).
+ * Единственная ответственность: получение списка категорий определенного типа из репозитория.
+ */
 class GetCategoriesByTypeUseCase @Inject constructor(
     private val categoryRepository: CategoryRepository
 ) {
