+++ app/src/main/java/yandex/school/project/domain/usecases/category/GetCategoriesUseCase.kt
@@ -4,6 +4,10 @@ import yandex.school.project.domain.entities.Category
 import yandex.school.project.domain.repositories.CategoryRepository
 import javax.inject.Inject
 
+/**
+ * Use case для получения списка всех категорий.
+ * Единственная ответственность: получение списка категорий из репозитория.
+ */
 class GetCategoriesUseCase @Inject constructor(
     private val categoryRepository: CategoryRepository
 ) {
