+++ app/src/main/java/yandex/school/project/presentation/screens/expenses/expenses/ExpensesViewModel.kt
@@ -1,4 +1,4 @@
-package yandex.school.project.presentation.screens.income
+package yandex.school.project.presentation.screens.expenses.expenses
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -6,22 +6,27 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import yandex.school.project.domain.entities.TransactionWithCategory
-import yandex.school.project.domain.usecases.category.GetCategoriesUseCase
 import yandex.school.project.domain.usecases.transaction.GetTransactionsByAccountUseCase
-import yandex.school.project.presentation.common.NetworkOperationHelper
+import yandex.school.project.domain.usecases.category.GetCategoriesUseCase
 import yandex.school.project.presentation.common.Result
+import yandex.school.project.presentation.common.NetworkOperationHelper
+import yandex.school.project.presentation.screens.expenses.ExpensesState
 import yandex.school.project.presentation.utils.CURRENCY_RUB
 import javax.inject.Inject
 
+/**
+ * ViewModel для экрана расходов, управляющий состоянием и загрузкой данных о расходах.
+ * Единственная ответственность: управление состоянием UI и загрузка данных о расходах с категориями.
+ */
 @HiltViewModel
-class IncomesViewModel @Inject constructor(
+class ExpensesViewModel @Inject constructor(
     private val getTransactionsByAccountUseCase: GetTransactionsByAccountUseCase,
     private val getCategoriesUseCase: GetCategoriesUseCase,
     private val networkHelper: NetworkOperationHelper
 ) : ViewModel() {
 
-    private val _uiState = MutableStateFlow<Result<IncomesState>>(Result.Loading)
-    val uiState: StateFlow<Result<IncomesState>> = _uiState
+    private val _uiState = MutableStateFlow<Result<ExpensesState>>(Result.Loading)
+    val uiState: StateFlow<Result<ExpensesState>> = _uiState
 
     fun loadTransactionsWithRetry(accountId: Int, maxRetries: Int = 3, delayMillis: Long = 2000) {
         networkHelper.executeWithRetry(
@@ -30,11 +35,11 @@ class IncomesViewModel @Inject constructor(
                 val allTransactions = getTransactionsByAccountUseCase(accountId)
                 val categories = getCategoriesUseCase()
                 
-                val incomeTransactions = allTransactions.filter { transaction ->
-                    transaction.type == yandex.school.project.domain.entities.TransactionType.INCOME
+                val expenseTransactions = allTransactions.filter { transaction ->
+                    transaction.type == yandex.school.project.domain.entities.TransactionType.EXPENSE
                 }
                 
-                val transactionsWithCategory = incomeTransactions.mapNotNull { transaction ->
+                val transactionsWithCategory = expenseTransactions.mapNotNull { transaction ->
                     val category = categories.find { it.id == transaction.categoryId }
                     category?.let { TransactionWithCategory(transaction, it) }
                 }.sortedByDescending {
@@ -42,20 +47,20 @@ class IncomesViewModel @Inject constructor(
                 }
                 
                 val total = transactionsWithCategory.sumOf { it.amount }
-                IncomesState(
+                ExpensesState(
                     transactions = transactionsWithCategory,
                     total = "${total.toInt()} $CURRENCY_RUB"
                 )
             },
-            onSuccess = { incomesState ->
-                _uiState.value = Result.Success(incomesState)
+            onSuccess = { expensesState ->
+                _uiState.value = Result.Success(expensesState)
             },
             onError = { errorMessage ->
                 _uiState.value = Result.Error(errorMessage)
             },
             maxRetries = maxRetries,
             delayMillis = delayMillis,
-            operationName = "загрузка доходов"
+            operationName = "загрузка расходов"
         )
     }
 }
