+++ app/src/main/java/yandex/school/project/presentation/screens/expenses/history/ExpensesHistoryViewModel.kt
@@ -1,7 +1,6 @@
-package yandex.school.project.presentation.screens.expenses
+package yandex.school.project.presentation.screens.expenses.history
 
 import android.os.Build
-import android.util.Log
 import androidx.annotation.RequiresApi
 import androidx.compose.runtime.*
 import androidx.lifecycle.ViewModel
