+++ app/src/main/java/yandex/school/project/presentation/screens/income/incomes/IncomesViewModel.kt
@@ -1,28 +1,32 @@
-package yandex.school.project.presentation.screens.expenses
+package yandex.school.project.presentation.screens.income.incomes
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.launch
 import yandex.school.project.domain.entities.TransactionWithCategory
-import yandex.school.project.domain.usecases.transaction.GetTransactionsByAccountUseCase
 import yandex.school.project.domain.usecases.category.GetCategoriesUseCase
-import yandex.school.project.presentation.common.Result
+import yandex.school.project.domain.usecases.transaction.GetTransactionsByAccountUseCase
 import yandex.school.project.presentation.common.NetworkOperationHelper
+import yandex.school.project.presentation.common.Result
+import yandex.school.project.presentation.screens.income.IncomesState
 import yandex.school.project.presentation.utils.CURRENCY_RUB
 import javax.inject.Inject
 
+/**
+ * ViewModel для экрана доходов, управляющий состоянием и загрузкой данных о доходах.
+ * Единственная ответственность: управление состоянием UI и загрузка данных о доходах с категориями.
+ */
 @HiltViewModel
-class ExpensesViewModel @Inject constructor(
+class IncomesViewModel @Inject constructor(
     private val getTransactionsByAccountUseCase: GetTransactionsByAccountUseCase,
     private val getCategoriesUseCase: GetCategoriesUseCase,
     private val networkHelper: NetworkOperationHelper
 ) : ViewModel() {
 
-    private val _uiState = MutableStateFlow<Result<ExpensesState>>(Result.Loading)
-    val uiState: StateFlow<Result<ExpensesState>> = _uiState
+    private val _uiState = MutableStateFlow<Result<IncomesState>>(Result.Loading)
+    val uiState: StateFlow<Result<IncomesState>> = _uiState
 
     fun loadTransactionsWithRetry(accountId: Int, maxRetries: Int = 3, delayMillis: Long = 2000) {
         networkHelper.executeWithRetry(
@@ -31,11 +35,11 @@ class ExpensesViewModel @Inject constructor(
                 val allTransactions = getTransactionsByAccountUseCase(accountId)
                 val categories = getCategoriesUseCase()
                 
-                val expenseTransactions = allTransactions.filter { transaction ->
-                    transaction.type == yandex.school.project.domain.entities.TransactionType.EXPENSE
+                val incomeTransactions = allTransactions.filter { transaction ->
+                    transaction.type == yandex.school.project.domain.entities.TransactionType.INCOME
                 }
                 
-                val transactionsWithCategory = expenseTransactions.mapNotNull { transaction ->
+                val transactionsWithCategory = incomeTransactions.mapNotNull { transaction ->
                     val category = categories.find { it.id == transaction.categoryId }
                     category?.let { TransactionWithCategory(transaction, it) }
                 }.sortedByDescending {
@@ -43,20 +47,20 @@ class ExpensesViewModel @Inject constructor(
                 }
                 
                 val total = transactionsWithCategory.sumOf { it.amount }
-                ExpensesState(
+                IncomesState(
                     transactions = transactionsWithCategory,
                     total = "${total.toInt()} $CURRENCY_RUB"
                 )
             },
-            onSuccess = { expensesState ->
-                _uiState.value = Result.Success(expensesState)
+            onSuccess = { incomesState ->
+                _uiState.value = Result.Success(incomesState)
             },
             onError = { errorMessage ->
                 _uiState.value = Result.Error(errorMessage)
             },
             maxRetries = maxRetries,
             delayMillis = delayMillis,
-            operationName = "загрузка расходов"
+            operationName = "загрузка доходов"
         )
     }
 }
