+++ app/src/main/java/yandex/school/project/presentation/screens/splash/SplashViewModel.kt
@@ -7,6 +7,10 @@ import kotlinx.coroutines.launch
 import yandex.school.project.domain.usecases.account.GetFirstAccountUseCase
 import javax.inject.Inject
 
+/**
+ * ViewModel для экрана загрузки, отвечающий за инициализацию приложения.
+ * Единственная ответственность: получение первого аккаунта для инициализации приложения.
+ */
 @HiltViewModel
 class SplashViewModel @Inject constructor(
     private val getFirstAccountUseCase: GetFirstAccountUseCase
