+++ app/src/main/java/yandex/school/project/presentation/utils/Constants.kt
@@ -1,5 +1,9 @@
 package yandex.school.project.presentation.utils
 
+/**
+ * Константы, используемые в presentation слое приложения.
+ * Единственная ответственность: хранение констант для UI и presentation логики.
+ */
 
 const val CURRENCY_RUB = "₽"
 const val BALANCE_ZERO = "0 $CURRENCY_RUB"
