+++ app/src/main/java/yandex/school/project/utils/NetworkErrorHandler.kt
@@ -11,7 +11,8 @@ import yandex.school.project.presentation.common.NetworkOperationHelper
 import yandex.school.project.presentation.utils.DEFAULT_OPERATION_NAME
 
 /**
- * Утилита для обработки сетевых ошибок
+ * Утилита для обработки сетевых ошибок.
+ * Единственная ответственность: классификация и обработка различных типов сетевых ошибок.
  */
 object NetworkErrorHandler {
